Extract risk calculation router
