+++ quality/refactor_health_report.md
@@ -8,7 +8,7 @@ This slice establishes the report-only enterprise quality baseline and progressi
 
 | Rank | Target | Evidence | Next action |
 | --- | --- | --- | --- |
-| 1 | API entry point | src/app/main.py has 305 lines and 5 app decorators | Extract routers and dependency providers |
+| 1 | API entry point | src/app/main.py has 256 lines and 4 app decorators | Extract routers and dependency providers |
 | 2 | Concentration module | Largest service and contract files are concentration-related | Split source resolution, issuer aggregation, and response assembly |
 | 3 | Risk calculation engine | Largest function is calculate_risk at 284 lines | Extract per-metric calculators behind stable service API |
